requirements.txt
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
cobble==0.1.3
idna==3.3
mammoth==1.4.19
python-slugify==6.1.2
requests==2.27.1
soupsieve==2.3.2.post1
text-unidecode==1.3
urllib3==1.26.9
